+++ CHANGELOG.md
@@ -6,3 +6,6 @@
 - Missing `<title>` on WebUI.
 - Infinite "node-in-node-list" notifications.
 - Some strange things which were happening with `Heartbeat`.
+
+### Added
+- Support for interacting with Cellframe Node CLI.